requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.0
SQLAlchemy==2.0.36
python-jose==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.27.2
email-validator==2.1.0.post1
